fixed routing

+++ app/Repositories/Interfaces/UserRepositoryInterface.php
@@ -7,5 +7,5 @@ use App\Models\User;
 interface UserRepositoryInterface{
     public function getUserById(string $userId): User;
     public function updateBalance(string $userId,string $type,int $balance,int $amount);
-    public function manualUserBalance(string $userId,int $amount);
+    public function manualUserBalance(string $userId,float $amount);
 }
